Remove all the stuff I won't need
I'm trying to be more pragmatic that last year. Debugging on this machine is extremely difficult, so I cannot implement any complicated algorithm on it.
If I really need a complex algorithm, I will have to implement it somewhere more viable and port it.
Anyway, let's cross that bridge when we get there!